+++ apps/backend/src/models/questionTypes/DateRange.ts
@@ -1,28 +1,10 @@
 /* eslint-disable quotes */
+import { dateRangeQuestionValidationSchema } from '@user-office-software/duo-validation';
 import { GraphQLError } from 'graphql';
-import * as Yup from 'yup';
 
 import { DateRangeConfig } from '../../resolvers/types/FieldConfig';
 import { DataType, QuestionTemplateRelation } from '../Template';
 import { Question } from './QuestionRegistry';
-
-//TODO add validation script to duo-validation
-export const dateRangeQuestionValidationSchema = (answer: any) => {
-  const config = answer.config;
-  let dateRangeSchema = Yup.array()
-    .of(
-      Yup.object({
-        from: Yup.date().required(),
-        to: Yup.date().required(),
-      })
-    )
-    .required();
-  if (config.required) {
-    dateRangeSchema = dateRangeSchema.min(1, 'A daterange is reqruired');
-  }
-
-  return Yup.object({ dateRanges: dateRangeSchema });
-};
 
 export const dateRangeDefinition: Question<DataType.DATE_RANGE_PICKER> = {
   dataType: DataType.DATE_RANGE_PICKER,
